Don't interrupt rating popups for the period of "daysUntilPrompt" on app update
Because all counters are reset on app update, no rating popups appear for daysUntilPrompt.
This patch changes this behaviour and avoids the counter reset. In the case the app was
updated and it's time to rate the app, it delays the rating by one day to give the user a chance
to notice the new features and fixed bugs.